requirements.txt
scikit-learn==0.19.1
librosa==0.6.0
matplotlib==2.2.2
sac==0.7.1
dlib==19.10.0
scikit-image==0.13.1
Keras==2.1.3
tensorflow==1.4.1
h5py==2.7.1
pyannote.metrics==1.7
